Update package names in `cli` module
