+++ cli/src/test/java/io/spine/examples/todolist/validator/NeitherCompletedNorDeletedValidatorTest.java
@@ -18,7 +18,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.spine3.examples.todolist.validator;
+package io.spine.examples.todolist.validator;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -37,34 +37,34 @@ class NeitherCompletedNorDeletedValidatorTest {
     private NeitherCompletedNorDeletedValidator validator;
 
     @BeforeEach
-    public void setUp() {
+    void setUp() {
         validator = new NeitherCompletedNorDeletedValidator();
     }
 
     @Test
     @DisplayName("pass the validation")
-    public void passValidation() {
+    void passValidation() {
         final boolean passed = validator.validate("Some input");
         assertTrue(passed);
     }
 
     @Test
     @DisplayName("not pass the validation when input is empty")
-    public void notPassValidationWhenInputIsEmpty() {
+    void notPassValidationWhenInputIsEmpty() {
         final boolean passed = validator.validate("");
         assertFalse(passed);
     }
 
     @Test
     @DisplayName("not pass the validation when input is null")
-    public void notPassValidationWhenInputIsNull() {
+    void notPassValidationWhenInputIsNull() {
         final boolean passed = validator.validate(null);
         assertFalse(passed);
     }
 
     @Test
     @DisplayName("return non null message when validation is failed")
-    public void getMessage() {
+    void getMessage() {
         validator.validate(null);
         assertNotNull(validator.getMessage());
     }
